added open npm
I added open to my node packages in order to open localhost automatically when I run a Powershell function. This is so I can run the learning program directly from the cli whenever I want to revise. I would likely remove this if it comes to ever putting this application in production.